+++ plutus-ledger-api/tests/serde.rs
@@ -158,4 +158,107 @@ mod serde_roundtrip_tests {
             }
         }
     }
+    mod v3 {
+        use super::from_to_json;
+        use plutus_ledger_api::generators::correct::v3::*;
+        use proptest::prelude::*;
+
+        proptest! {
+            #[test]
+            fn test_cold_committee_credential(val in arb_cold_committee_credential()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_hot_committee_credential(val in arb_hot_committee_credential()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_d_rep_credential(val in arb_d_rep_credential()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_delegatee(val in arb_delegatee()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_tx_cert(val in arb_tx_cert()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_voter(val in arb_voter()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_vote(val in arb_vote()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_governance_action_id(val in arb_governance_action_id()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_committee(val in arb_committee()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_rational(val in arb_rational()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_constitution(val in arb_constitution()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_protocol_version(val in arb_protocol_version()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_change_parameters(val in arb_change_parameters()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+
+            #[test]
+            fn test_governance_action(val in arb_governance_action()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_protocol_procedure(val in arb_protocol_procedure()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_script_purpose(val in arb_script_purpose()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_script_info(val in arb_script_info()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_transaction_info(val in arb_transaction_info()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+
+            #[test]
+            fn test_script_context(val in arb_script_context()) {
+                assert_eq!(val, from_to_json(&val)?)
+            }
+        }
+    }
 }