+++ plutus-ledger-api/CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog](https://keepachangelog.com/en/1.1.0).
 
 ### Added
 
+- Added v3 plutus ledger types ([#57](https://github.com/mlabs-haskell/plutus-ledger-api-rust/pull/57))
 - Added the ability to derive `IsPlutusData` instances ([#56](https://github.com/mlabs-haskell/plutus-ledger-api-rust/pull/56))
 
 ### Changed